# openapi.json
{
  "components": {
    "schemas": {
      "AddressStructure": {
        "additionalProperties": false,
        "description": "Structured address fields using vworld-compatible names.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "level0": {
            "default": "대한민국",
            "minLength": 1,
            "title": "Level0",
            "type": "string"
          },
          "level1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level1"
          },
          "level2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level2"
          },
          "level3": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level3"
          },
          "level4A": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level4A"
          },
          "level4AC": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level4Ac"
          },
          "level4L": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level4L"
          },
          "level4LC": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level4Lc"
          },
          "level5": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level5"
          }
        },
        "title": "AddressStructure",
        "type": "object"
      },
      "AddressType": {
        "enum": [
          "road",
          "parcel"
        ],
        "type": "string"
      },
...
      "BBox": {
        "additionalProperties": false,
        "properties": {
          "max_x": {
            "title": "Max X",
            "type": "number"
          },
          "max_y": {
            "title": "Max Y",
            "type": "number"
          },
          "min_x": {
            "title": "Min X",
            "type": "number"
          },
          "min_y": {
            "title": "Min Y",
            "type": "number"
          }
        },
        "required": [
          "min_x",
          "min_y",
          "max_x",
          "max_y"
        ],
        "title": "BBox",
        "type": "object"
      },
...
      "CRS": {
        "description": "Coordinate reference system normalized as EPSG:XXXX",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Point": {
        "additionalProperties": false,
        "description": "Coordinate point. External interfaces use ``x=lon`` and ``y=lat``.",
        "properties": {
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "Point",
        "type": "object"
      },
...
      "ResultSource": {
        "enum": [
          "local",
          "api_juso",
          "api_vworld",
          "cache"
        ],
        "type": "string"
      },
      "ReverseExtension": {
        "additionalProperties": false,
        "properties": {
          "sppn_makarea": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SppnMakareaContext"
            },
            "title": "Sppn Makarea",
            "type": "array"
          }
        },
        "title": "ReverseExtension",
        "type": "object"
      },
      "ReverseInput": {
        "additionalProperties": false,
        "properties": {
          "crs": {
            "$ref": "#/components/schemas/CRS",
            "default": "EPSG:4326"
          },
          "point": {
            "$ref": "#/components/schemas/Point"
          },
          "radius_m": {
            "default": 200,
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Radius M",
            "type": "integer"
          },
          "type": {
            "default": "both",
            "enum": [
              "both",
              "road",
              "parcel"
            ],
            "title": "Type",
            "type": "string"
          },
          "zipcode": {
            "default": true,
            "title": "Zipcode",
            "type": "boolean"
          }
        },
        "required": [
          "point"
        ],
        "title": "ReverseInput",
        "type": "object"
      },
      "ReverseResponse": {
        "additionalProperties": false,
        "properties": {
          "input": {
            "$ref": "#/components/schemas/ReverseInput"
          },
          "result": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ReverseResultItem"
            },
            "title": "Result",
            "type": "array"
          },
          "service": {
            "$ref": "#/components/schemas/ServiceMeta"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "x_extension": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReverseExtension"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "service",
          "status",
          "input"
        ],
        "title": "ReverseResponse",
        "type": "object"
      },
      "ReverseResultItem": {
        "additionalProperties": false,
        "properties": {
          "distance_m": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "point": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Point"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "$ref": "#/components/schemas/ResultSource",
            "default": "local"
          },
          "structure": {
            "$ref": "#/components/schemas/AddressStructure"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AddressType"
          },
          "zip_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ZipSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "zipcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zipcode"
          }
        },
        "required": [
          "type",
          "text",
          "structure"
        ],
        "title": "ReverseResultItem",
        "type": "object"
      },
...
      "SearchInput": {
        "additionalProperties": false,
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "crs": {
            "$ref": "#/components/schemas/CRS",
            "default": "EPSG:4326"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "query": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "size": {
            "default": 10,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Size",
            "type": "integer"
          },
          "type": {
            "default": "address",
            "enum": [
              "address",
              "place",
              "district",
              "road"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchInput",
        "type": "object"
      },
      "SearchResponse": {
        "additionalProperties": false,
        "properties": {
          "input": {
            "$ref": "#/components/schemas/SearchInput"
          },
          "result": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "title": "Result",
            "type": "array"
          },
          "service": {
            "$ref": "#/components/schemas/ServiceMeta"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "total": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "service",
          "status",
          "input"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "SearchResultItem": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "point": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Point"
              },
              {
                "type": "null"
              }
            ]
          },
          "score": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "source": {
            "$ref": "#/components/schemas/ResultSource",
            "default": "local"
          },
          "structure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AddressStructure"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "enum": [
              "address",
              "place",
              "district",
              "road"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title"
        ],
        "title": "SearchResultItem",
        "type": "object"
      },
...
      "ServiceMeta": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          },
          "version": {
            "default": "2.0",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "operation"
        ],
        "title": "ServiceMeta",
        "type": "object"
      },
...
      "SppnMakareaContext": {
        "additionalProperties": false,
        "properties": {
          "area_m2": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area M2"
          },
          "makarea_id": {
            "title": "Makarea Id",
            "type": "string"
          },
          "makarea_nm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Makarea Nm"
          },
          "mvm_res_cd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mvm Res Cd"
          },
          "ntfc_de": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ntfc De"
          },
          "ntfc_yn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ntfc Yn"
          },
          "sig_cd": {
            "title": "Sig Cd",
            "type": "string"
          },
          "source_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source File"
          },
          "source_yyyymm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Yyyymm"
          }
        },
        "required": [
          "sig_cd",
          "makarea_id"
        ],
        "title": "SppnMakareaContext",
        "type": "object"
      },
      "Status": {
        "enum": [
          "OK",
          "NOT_FOUND",
          "ERROR"
        ],
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ZipSource": {
        "enum": [
          "building_bsi_zon_no",
          "bulk_delivery",
          "kodis_bas_within",
          "kodis_bas_centroid",
          "pobox"
        ],
        "title": "ZipSource",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "kor-travel-geo",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/address/reverse": {
      "get": {
        "operationId": "reverse_geocode_v1_address_reverse_get",
        "parameters": [
          {
            "in": "query",
            "name": "x",
            "required": true,
            "schema": {
              "title": "X",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "y",
            "required": true,
            "schema": {
              "title": "Y",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "crs",
            "required": false,
            "schema": {
              "default": "EPSG:4326",
              "title": "Crs",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "both",
              "enum": [
                "both",
                "road",
                "parcel"
              ],
              "title": "Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "zipcode",
            "required": false,
            "schema": {
              "default": true,
              "title": "Zipcode",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "radius_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Radius M"
            }
          },
          {
            "in": "query",
            "name": "sig_cd",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(\\d{2}|\\d{5})$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sig Cd"
            }
          },
          {
            "in": "query",
            "name": "bjd_cd",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(\\d{8}|\\d{10})$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bjd Cd"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReverseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reverse Geocode",
        "tags": [
          "address"
        ]
      }
    },
    "/v1/address/search": {
      "get": {
        "operationId": "search_v1_address_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "address",
              "enum": [
                "address",
                "place",
                "district",
                "road"
              ],
              "title": "Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sig_cd",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(\\d{2}|\\d{5})$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sig Cd"
            }
          },
          {
            "in": "query",
            "name": "bjd_cd",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(\\d{8}|\\d{10})$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bjd Cd"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}